Logging: hexdump fixes

+++ Kernel/Logging.h
@@ -3,12 +3,12 @@
 
 #include <stddef.h>
 
-#define DEBUG 0
-#define SUCCESS 1
-#define INFO 2
-#define WARNING 3
-#define ERROR 4
-#define PANIC 5
+#define DEBUG		0
+#define SUCCESS		1
+#define INFO		2
+#define WARNING		3
+#define ERROR		4
+#define PANIC		5
 
 /// @brief Log a formatted string
 /// @param logLevel SUCCESS, INFO, WARNING, ERROR... See `Logging.h` for values
@@ -17,8 +17,12 @@
 /// @param ... formatting arguments (printf syntax)
 void log(int logLevel, const char* moduleName, const char* logFmtStr, ...);
 
-/// @brief Dump the memory (in hexadecimal) in the console
-void hexdump(int logLevel, const char* moduleName, void* addr , size_t n);
+/// @brief Dump the memory (in hexadecimal) to the console
+/// @param logLevel SUCCESS, INFO, WARNING, ERROR... See `Logging.h` for values
+/// @param moduleName Nullable string of the module from which log is emitted
+/// @param addr[n] Memory address of the data to dump
+/// @param n Number of bytes from `addr` to dump
+void hexdump(int logLevel, const char* moduleName, void* addr , int n);
 
 /// @brief Alias for `log`, with default arguments.
 /// Handy for writing quick logs while debugging
